+++ sentry-ruby/lib/sentry/envelope/item.rb
@@ -3,7 +3,6 @@
 module Sentry
   # @api private
   class Envelope::Item
-    STACKTRACE_FRAME_LIMIT_ON_OVERSIZED_PAYLOAD = 500
     MAX_SERIALIZED_PAYLOAD_SIZE = 1024 * 1000
 
     SIZE_LIMITS = Hash.new(MAX_SERIALIZED_PAYLOAD_SIZE).update(
@@ -45,11 +44,6 @@ module Sentry
         result = to_s
       end
 
-      if result.bytesize > size_limit
-        reduce_stacktrace!
-        result = to_s
-      end
-
       [result, result.bytesize > size_limit]
     end
 
@@ -66,22 +60,6 @@ module Sentry
         payload.delete(:breadcrumbs)
       elsif payload.key?("breadcrumbs")
         payload.delete("breadcrumbs")
-      end
-    end
-
-    def reduce_stacktrace!
-      if exceptions = payload.dig(:exception, :values) || payload.dig("exception", "values")
-        exceptions.each do |exception|
-          # in most cases there is only one exception (2 or 3 when have multiple causes), so we won't loop through this double condition much
-          traces = exception.dig(:stacktrace, :frames) || exception.dig("stacktrace", "frames")
-
-          if traces && traces.size > STACKTRACE_FRAME_LIMIT_ON_OVERSIZED_PAYLOAD
-            size_on_both_ends = STACKTRACE_FRAME_LIMIT_ON_OVERSIZED_PAYLOAD / 2
-            traces.replace(
-              traces[0..(size_on_both_ends - 1)] + traces[-size_on_both_ends..-1],
-            )
-          end
-        end
       end
     end
   end